+++ src/design/plone/contenttypes/tests/test_ct_folder.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+
+from design.plone.contenttypes.testing import (
+    DESIGN_PLONE_CONTENTTYPES_API_FUNCTIONAL_TESTING,
+)
+from plone import api
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
+from plone.app.testing import SITE_OWNER_NAME
+from plone.app.testing import SITE_OWNER_PASSWORD
+from plone.restapi.testing import RelativeSession
+
+import unittest
+
+
+class TestFolderSchema(unittest.TestCase):
+    layer = DESIGN_PLONE_CONTENTTYPES_API_FUNCTIONAL_TESTING
+
+    def setUp(self):
+        self.app = self.layer["app"]
+        self.portal = self.layer["portal"]
+        self.request = self.layer["request"]
+        self.portal_url = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ["Manager"])
+
+        self.api_session = RelativeSession(self.portal_url)
+        self.api_session.headers.update({"Accept": "application/json"})
+        self.api_session.auth = (SITE_OWNER_NAME, SITE_OWNER_PASSWORD)
+
+    def tearDown(self):
+        self.api_session.close()
+
+    def test_behaviors_enabled_for_folder(self):
+        portal_types = api.portal.get_tool(name="portal_types")
+        self.assertEqual(
+            portal_types["Folder"].behaviors,
+            (
+                "plone.dublincore",
+                "plone.namefromtitle",
+                "plone.allowdiscussion",
+                "plone.excludefromnavigation",
+                "plone.shortname",
+                "plone.constraintypes",
+                "plone.relateditems",
+                "plone.nextprevioustoggle",
+                "design.plone.contenttypes.behavior.exclude_from_search",
+            ),
+        )