+++ CHANGES.rst
@@ -12,6 +12,13 @@ Changelog
   [cekk]
 - Improve types test for their schema, required fields, fieldsets.
   [cekk]
+- Add *exclude_from_search* indexer and behavior, and enable for Document and Folder.
+  [cekk]
+- Add custom adapter for IZCatalogCompatibleQuery to force all anonymous @search calls to skip items excluded from search.
+  [cekk]
+- Set *exclude_from_search* to True in all Documents/Folders automatically created in createSubfolders event handler,
+  and add an upgrade-step that fix already created ones.
+  [cekk]
 
 6.1.14 (2024-02-20)
 -------------------